+++ bubblemachine-vite-react/src/components/layout/BubbleBox.jsx
