+++ bubblemachine-vite-react/src/components/layout/Footer.jsx
