+++ bubblemachine-vite-react/src/components/layout/Header.jsx
@@ -1,7 +1,6 @@
 import React from "react";
 import { AppBar, Toolbar, Typography, Box } from "@mui/material";
 import logo from "../assets/BubbleMachine_Transparent.png"; // Adjust the path if necessary
-
 const Header = () => {
   return (
     <AppBar position="static">
