+++ bubblemachine-vite-react/src/components/waveform/Progressbar.jsx
