+++ bubblemachine-vite-react/src/components/waveform/WaveformControls.jsx
