+++ bubblemachine-vite-react/src/helpers/utils.jsx
