+++ bubblemachine-vite-react/src/pages/Layout.jsx
@@ -1,7 +1,7 @@
 // src/pages/Layout.jsx
 import React from "react";
-import Header from "../components/Header";
-import Footer from "../components/Footer";
+import Header from "../components/layout/Header";
+import Footer from "../components/layout/Footer";
 import { Outlet, Link } from "react-router-dom";
 
 export default function Layout() {
